Add new classes for OOP
