--- src/main/duke.txt
@@ -1,4 +1,0 @@
-todo wash dishes
-true
-todo 2103
-false
