+++ src/main/java/Duke.java
@@ -1,5 +1,11 @@
+import ExceptionHandling.DukeException;
+import Tasks.Deadline;
+import Tasks.Event;
+import Tasks.Task;
+import Tasks.Todo;
+
 import java.io.File;
-import java.io.FileWriter;   // Import the FileWriter class
+import java.io.FileWriter;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
