+++ src/main/java/ExceptionHandling/DukeException.java
@@ -1,3 +1,5 @@
+package ExceptionHandling;
+
 public class DukeException extends Exception{
     public DukeException(String message) {
         super(message);
