+++ src/main/java/Parser/Parser.java
@@ -1,0 +1,5 @@
+package Parser;
+
+public class Parser {
+    //pass in the chat bot and read command function
+}
