+++ src/main/java/Storage/Storage.java
@@ -1,0 +1,5 @@
+package Storage;
+
+public class Storage {
+    //move the save/load here
+}
