+++ src/main/java/Tasks/Event.java
@@ -1,4 +1,8 @@
-public class Event extends Task{
+package Tasks;
+
+import Tasks.Task;
+
+public class Event extends Task {
     private String from;
     private String to;
     public Event(String description, String from, String to) {
